requirements.txt
allure-pytest==2.8.16
coverage==5.2.1
numpy==1.18.1
openpyxl==3.0.3
pandas==1.0.1
Pillow==7.0.0
PyAutoGUI==0.9.48
PyAutoIt==0.6.0
pyjavaproperties==0.7
pylint==2.5.3
pytest==5.4.3
pytest-xdist==1.34.0
selenium==3.141.0
Sphinx==3.1.2
xlrd==1.2.0
